Reorganize!
Restructured a bit and fixed a super annoying "expression not defined" issue with existing Alpine stuff.

1. The cookie popup is now its own partial. This makes far more sense because it's not REALLY part of the whole Google Analytics thing. It just hooks into part of it.

2. Google Analytics is a Volt component now, which means Alpine is pulled in directly. There's no need to load Alpine separately anymore.

3. A separate Alpine global file actually does exist now, for if it's ever needed. It shouldn't be, but it's there just in case.

4. All the old GA and popup stuff is removed since both are set up different ways now.

5. The listener for that custom event is now collocated with the actual handler. Don't know why I didn't do this before, it makes FAR more sense this way.
