+++ vite.config.ts
@@ -10,7 +10,11 @@ export default defineConfig(({mode}) => {
 	return {
 		plugins: [
 			laravel({
-				input: ['resources/css/app.css', 'resources/js/app.ts'],
+				input: [
+					'resources/css/app.css',
+					'resources/js/alpine-init.ts',
+					'resources/js/app.ts',
+				],
 				refresh: true,
 			}),
 			vuePlugin(),
